requirements.txt
web3==6.20.4
python-dotenv==1.0.1
aiohttp==3.9.3
APScheduler==3.10.4
eth-account==0.11.0
eth-abi==5.0.1
requests==2.31.0
